Implement the BufferList class
This is used by the editor to track all open buffers. A BufferList
instance only allows one buffer to be added to the list for a given
filepath, so the same file cannot be edited in two different buffers at
the same time.

The 'open' command was also modified so that opening the
same file twice merely restored an existing buffer, and opening the same
file in multiple panes allowed them to share a buffer without being
linked.